ramips: use partition table on Netgear NAND devices
There have been multiple reports of the partition table on these devices
being incorrect when there are bad blocks. This change should fix this by
parsing the partition table that is stored in the SC_PART_MAP partition.

This is essentially the same code as proposed in the original pull request
for Netgear R6350 support by NOGUCHI Hiroshi <drvlabo@gmail.com>:
https://github.com/openwrt/openwrt/pull/1318

At the time, it was rejected in favor of fixed partition offsets, as it
did not seem to work correctly. However, this was due the NAND driver
transparently shifting pages to hide bad blocks, which was fixed in
commit 527832e54bf3bc4d699a145ae66f34230246f0a9.

Signed-off-by: Jan Hoffmann <jan@3e8.eu>