ltq-vectoring: add driver
This driver provides a callback for sending vectoring error samples.

It is based on the open source code released for D-Link DWR-966.
Changes were made to support more recent kernel versions and the used
interface was changed from ptm0 to dsl0.

NOTE: this is currently broken as the PTM driver does not support
concurrent calls to ndo_start_xmit.

Signed-off-by: Jan Hoffmann <jan@3e8.eu>